#25 Deploy the Bug Hunt bot
* Move the Bug Hunt bot to the separate module
* Add the clearing procedure
