+++ bug-hunt-bot/pom.xml
@@ -31,7 +31,7 @@
                     <archive>
                         <manifest>
                             <addClasspath>true</addClasspath>
-                            <mainClass>press.lis.greeb.bug_hunt.BugHuntBotAppKt</mainClass>
+                            <mainClass>press.lis.greeb.bug_hunt.BugHuntBotKt</mainClass>
                             <classpathPrefix>lib/</classpathPrefix>
                         </manifest>
                     </archive>
